Refactor PrivateDataStore (#1715)
* refactor pvtdatastore


* mv oldpvtdata commit to separate file
