+++ core/ledger/pvtdatastorage/store.go
@@ -107,15 +107,6 @@ type storeEntries struct {
 // and is stored as the value of lastUpdatedOldBlocksKey (defined in kv_encoding.go)
 type lastUpdatedOldBlocksList []uint64
 
-type entriesForPvtDataOfOldBlocks struct {
-	// for each <ns, coll, blkNum, txNum>, store the dataEntry, i.e., pvtData
-	dataEntries map[dataKey]*rwset.CollectionPvtReadWriteSet
-	// store the retrieved (& updated) expiryData in expiryEntries
-	expiryEntries map[expiryKey]*ExpiryData
-	// for each <ns, coll, blkNum>, store the retrieved (& updated) bitmap in the missingDataEntries
-	missingDataEntries map[nsCollBlk]*bitset.BitSet
-}
-
 //////// Provider functions  /////////////
 //////////////////////////////////////////
 
@@ -264,229 +255,6 @@ func (s *Store) Commit(blockNum uint64, pvtData []*ledger.TxPvtData, missingPvtD
 	logger.Debugf("Committed private data for block [%d]", committingBlockNum)
 	s.performPurgeIfScheduled(committingBlockNum)
 	return nil
-}
-
-// CommitPvtDataOfOldBlocks commits the pvtData (i.e., previously missing data) of old blocks.
-// The parameter `blocksPvtData` refers a list of old block's pvtdata which are missing in the pvtstore.
-// Given a list of old block's pvtData, `CommitPvtDataOfOldBlocks` performs the following four
-// operations
-// (1) construct dataEntries for all pvtData
-// (2) construct update entries (i.e., dataEntries, expiryEntries, missingDataEntries)
-//     from the above created data entries
-// (3) create a db update batch from the update entries
-// (4) commit the update batch to the pvtStore
-func (s *Store) CommitPvtDataOfOldBlocks(blocksPvtData map[uint64][]*ledger.TxPvtData) error {
-	if s.isLastUpdatedOldBlocksSet {
-		return &ErrIllegalCall{`The lastUpdatedOldBlocksList is set. It means that the
-		stateDB may not be in sync with the pvtStore`}
-	}
-
-	logger.Debugf("Constructing pvtdatastore entries for pvtData of [%d] old blocks", len(blocksPvtData))
-	updateEntries, err := s.constructUpdateEntries(blocksPvtData)
-	if err != nil {
-		return err
-	}
-
-	logger.Debug("Constructing update batch from pvtdatastore entries")
-	batch, err := s.constructUpdateBatch(updateEntries)
-	if err != nil {
-		return err
-	}
-
-	logger.Debug("Committing the update batch to pvtdatastore")
-	return s.commitBatch(batch)
-}
-
-func (s *Store) constructUpdateEntries(blocksPvtData map[uint64][]*ledger.TxPvtData) (*entriesForPvtDataOfOldBlocks, error) {
-	var dataEntries []*dataEntry
-	for blkNum, pvtData := range blocksPvtData {
-		dataEntries = append(dataEntries, prepareDataEntries(blkNum, pvtData)...)
-	}
-
-	updateEntries := &entriesForPvtDataOfOldBlocks{
-		dataEntries:        make(map[dataKey]*rwset.CollectionPvtReadWriteSet),
-		expiryEntries:      make(map[expiryKey]*ExpiryData),
-		missingDataEntries: make(map[nsCollBlk]*bitset.BitSet)}
-
-	// for each data entry, first, get the expiryData and missingData from the pvtStore.
-	// Second, update the expiryData and missingData as per the data entry. Finally, add
-	// the data entry along with the updated expiryData and missingData to the update entries
-	for _, dataEntry := range dataEntries {
-		// get the expiryBlk number to construct the expiryKey
-		expiryKey, err := s.constructExpiryKeyFromDataEntry(dataEntry)
-		if err != nil {
-			return nil, err
-		}
-
-		// get the existing expiryData entry
-		var expiryData *ExpiryData
-		if !neverExpires(expiryKey.expiringBlk) {
-			if expiryData, err = s.getExpiryDataFromUpdateEntriesOrStore(updateEntries, expiryKey); err != nil {
-				return nil, err
-			}
-			if expiryData == nil {
-				// data entry is already expired
-				// and purged (a rare scenario)
-				continue
-			}
-		}
-
-		// get the existing missingData entry
-		var missingData *bitset.BitSet
-		nsCollBlk := dataEntry.key.nsCollBlk
-		if missingData, err = s.getMissingDataFromUpdateEntriesOrStore(updateEntries, nsCollBlk); err != nil {
-			return nil, err
-		}
-		if missingData == nil {
-			// data entry is already expired
-			// and purged (a rare scenario)
-			continue
-		}
-
-		updateEntries.addDataEntry(dataEntry)
-		if expiryData != nil { // would be nil for the never expiring entry
-			expiryEntry := &expiryEntry{&expiryKey, expiryData}
-			updateEntries.updateAndAddExpiryEntry(expiryEntry, dataEntry.key)
-		}
-		updateEntries.updateAndAddMissingDataEntry(missingData, dataEntry.key)
-	}
-	return updateEntries, nil
-}
-
-func (s *Store) constructExpiryKeyFromDataEntry(dataEntry *dataEntry) (expiryKey, error) {
-	// get the expiryBlk number to construct the expiryKey
-	nsCollBlk := dataEntry.key.nsCollBlk
-	expiringBlk, err := s.btlPolicy.GetExpiringBlock(nsCollBlk.ns, nsCollBlk.coll, nsCollBlk.blkNum)
-	if err != nil {
-		return expiryKey{}, err
-	}
-	return expiryKey{expiringBlk, nsCollBlk.blkNum}, nil
-}
-
-func (s *Store) getExpiryDataFromUpdateEntriesOrStore(updateEntries *entriesForPvtDataOfOldBlocks, expiryKey expiryKey) (*ExpiryData, error) {
-	expiryData, ok := updateEntries.expiryEntries[expiryKey]
-	if !ok {
-		var err error
-		expiryData, err = s.getExpiryDataOfExpiryKey(&expiryKey)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return expiryData, nil
-}
-
-func (s *Store) getMissingDataFromUpdateEntriesOrStore(updateEntries *entriesForPvtDataOfOldBlocks, nsCollBlk nsCollBlk) (*bitset.BitSet, error) {
-	missingData, ok := updateEntries.missingDataEntries[nsCollBlk]
-	if !ok {
-		var err error
-		missingDataKey := &missingDataKey{nsCollBlk, true}
-		missingData, err = s.getBitmapOfMissingDataKey(missingDataKey)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return missingData, nil
-}
-
-func (updateEntries *entriesForPvtDataOfOldBlocks) addDataEntry(dataEntry *dataEntry) {
-	dataKey := dataKey{dataEntry.key.nsCollBlk, dataEntry.key.txNum}
-	updateEntries.dataEntries[dataKey] = dataEntry.value
-}
-
-func (updateEntries *entriesForPvtDataOfOldBlocks) updateAndAddExpiryEntry(expiryEntry *expiryEntry, dataKey *dataKey) {
-	txNum := dataKey.txNum
-	nsCollBlk := dataKey.nsCollBlk
-	// update
-	expiryEntry.value.addPresentData(nsCollBlk.ns, nsCollBlk.coll, txNum)
-	// we cannot delete entries from MissingDataMap as
-	// we keep only one entry per missing <ns-col>
-	// irrespective of the number of txNum.
-
-	// add
-	expiryKey := expiryKey{expiryEntry.key.expiringBlk, expiryEntry.key.committingBlk}
-	updateEntries.expiryEntries[expiryKey] = expiryEntry.value
-}
-
-func (updateEntries *entriesForPvtDataOfOldBlocks) updateAndAddMissingDataEntry(missingData *bitset.BitSet, dataKey *dataKey) {
-	txNum := dataKey.txNum
-	nsCollBlk := dataKey.nsCollBlk
-	// update
-	missingData.Clear(uint(txNum))
-	// add
-	updateEntries.missingDataEntries[nsCollBlk] = missingData
-}
-
-func (s *Store) constructUpdateBatch(updateEntries *entriesForPvtDataOfOldBlocks) (*leveldbhelper.UpdateBatch, error) {
-	batch := s.db.NewUpdateBatch()
-
-	// add the following four types of entries to the update batch: (1) new data entries
-	// (i.e., pvtData), (2) updated expiry entries, (3) updated missing data entries, and
-	// (4) updated block list
-
-	// (1) add new data entries to the batch
-	if err := addNewDataEntriesToUpdateBatch(batch, updateEntries); err != nil {
-		return nil, err
-	}
-
-	// (2) add updated expiryEntry to the batch
-	if err := addUpdatedExpiryEntriesToUpdateBatch(batch, updateEntries); err != nil {
-		return nil, err
-	}
-
-	// (3) add updated missingData to the batch
-	if err := addUpdatedMissingDataEntriesToUpdateBatch(batch, updateEntries); err != nil {
-		return nil, err
-	}
-
-	return batch, nil
-}
-
-func addNewDataEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch, entries *entriesForPvtDataOfOldBlocks) error {
-	var keyBytes, valBytes []byte
-	var err error
-	for dataKey, pvtData := range entries.dataEntries {
-		keyBytes = encodeDataKey(&dataKey)
-		if valBytes, err = encodeDataValue(pvtData); err != nil {
-			return err
-		}
-		batch.Put(keyBytes, valBytes)
-	}
-	return nil
-}
-
-func addUpdatedExpiryEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch, entries *entriesForPvtDataOfOldBlocks) error {
-	var keyBytes, valBytes []byte
-	var err error
-	for expiryKey, expiryData := range entries.expiryEntries {
-		keyBytes = encodeExpiryKey(&expiryKey)
-		if valBytes, err = encodeExpiryValue(expiryData); err != nil {
-			return err
-		}
-		batch.Put(keyBytes, valBytes)
-	}
-	return nil
-}
-
-func addUpdatedMissingDataEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch, entries *entriesForPvtDataOfOldBlocks) error {
-	var keyBytes, valBytes []byte
-	var err error
-	for nsCollBlk, missingData := range entries.missingDataEntries {
-		keyBytes = encodeMissingDataKey(&missingDataKey{nsCollBlk, true})
-		// if the missingData is empty, we need to delete the missingDataKey
-		if missingData.None() {
-			batch.Delete(keyBytes)
-			continue
-		}
-		if valBytes, err = encodeMissingDataValue(missingData); err != nil {
-			return err
-		}
-		batch.Put(keyBytes, valBytes)
-	}
-	return nil
-}
-
-func (s *Store) commitBatch(batch *leveldbhelper.UpdateBatch) error {
-	return s.db.WriteBatch(batch, true)
 }
 
 // GetLastUpdatedOldBlocksPvtData returns the pvtdata of blocks listed in `lastUpdatedOldBlocksList`
